rename use-counter to react-hooks
